Start reworking leaderboard to return a promise
